# src/api/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ ],
  "info" : {
    "description" : "Simple service for execute commands on remote host\n",
    "version" : "1.0.0",
    "title" : "ezci",
    "contact" : {
      "email" : "romgalkin23@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "projects",
    "description" : "Methods for managing your created projects."
  }, {
    "name" : "scripts",
    "description" : "Methods for manipulating with project scripts"
  }, {
    "name" : "executions",
    "description" : "Methods for manipulating with script executions"
  } ],
  "paths" : {
    "/project" : {
      "post" : {
        "tags" : [ "projects" ],
        "summary" : "Add a new project to your ci",
        "operationId" : "addProject",
        "responses" : {
          "400" : {
            "description" : "Invalid input"
          },
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Project"
        }
      },
      "get" : {
        "tags" : [ "projects" ],
        "summary" : "Fetch all projects",
        "operationId" : "getProjects",
        "responses" : {
          "200" : {
            "description" : "Successfull operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/{uid}" : {
      "get" : {
        "tags" : [ "projects" ],
        "summary" : "Fetch project by its uid",
        "operationId" : "getProject",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "description" : "uid of project to update",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Invalid uid supplied"
          },
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "projects" ],
        "summary" : "Update an existing project",
        "operationId" : "updateProject",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "description" : "uid of project to update",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Invalid uid supplied"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Project"
        }
      },
      "delete" : {
        "tags" : [ "projects" ],
        "summary" : "Delete an existing project",
        "operationId" : "deleteProject",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "description" : "uid of project to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Invalid uid supplied"
          },
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/script" : {
      "post" : {
        "tags" : [ "scripts" ],
        "summary" : "Add new script to project",
        "operationId" : "addScript",
        "parameters" : [ {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid not found"
          },
          "400" : {
            "description" : "Validation error"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Script"
                }
              }
            }
          }
        },
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Script"
        }
      },
      "get" : {
        "tags" : [ "scripts" ],
        "summary" : "Fetch all project scripts",
        "operationId" : "getScripts",
        "parameters" : [ {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid not found"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Script"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/script/{id}" : {
      "get" : {
        "tags" : [ "scripts" ],
        "summary" : "Fetch an existing script by its id",
        "operationId" : "getScriptById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Script id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid or script with supplied name not found"
          },
          "400" : {
            "description" : "Validation error"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Script"
                }
              }
            }
          }
        }
      }
    },
    "/script/{name}" : {
      "put" : {
        "tags" : [ "scripts" ],
        "summary" : "Update an existing script by its name",
        "operationId" : "updateScript",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Script name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid or script with supplied name not found"
          },
          "400" : {
            "description" : "Validation error"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Script"
                }
              }
            }
          }
        },
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Script"
        }
      },
      "delete" : {
        "tags" : [ "scripts" ],
        "summary" : "Update an existing script by its name",
        "operationId" : "deleteScript",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Script name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid or script with supplied name not found"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Script"
                }
              }
            }
          }
        }
      },
      "get" : {
        "tags" : [ "scripts" ],
        "summary" : "Fetch an existing script by its name",
        "operationId" : "getScript",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Script name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid or script with supplied name not found"
          },
          "400" : {
            "description" : "Validation error"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Script"
                }
              }
            }
          }
        }
      }
    },
    "/script/{name}/executions" : {
      "get" : {
        "tags" : [ "scripts" ],
        "summary" : "Fetch an existing script by its name",
        "operationId" : "getScriptExecutions",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Script name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid or script with supplied name not found"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Execution"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/script/{name}/execute" : {
      "post" : {
        "tags" : [ "scripts" ],
        "summary" : "Execute script on remote host",
        "operationId" : "execute",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Script name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "uid",
          "in" : "query",
          "description" : "Project uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Project with supplied uid or script with supplied name not found"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Execution"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "build" : {
                    "type" : "string",
                    "format" : "binary",
                    "description" : "Optional archive with build"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/execution/{id}" : {
      "get" : {
        "tags" : [ "executions" ],
        "summary" : "Fetch an exicution of script by its id",
        "operationId" : "getExecution",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Execution id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Execution with supplied id not found"
          },
          "200" : {
            "description" : "Succesful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Execution"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Project" : {
        "type" : "object",
        "properties" : {
          "uid" : {
            "type" : "string",
            "description" : "Unique token of project for manipulation with"
          },
          "name" : {
            "type" : "string",
            "description" : "Name of project"
          },
          "cwd" : {
            "type" : "string",
            "description" : "Absolute project path on server"
          },
          "host" : {
            "type" : "string",
            "description" : "Ip address of host machine"
          },
          "user" : {
            "type" : "string",
            "description" : "System user which will execute commands"
          }
        }
      },
      "Script" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Name of script"
          },
          "script" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "description" : "Formally lines of script to execute"
            }
          }
        }
      },
      "Execution" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "number"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "outs" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "code" : {
                  "type" : "string"
                },
                "stdout" : {
                  "type" : "string"
                },
                "stderr" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "requestBodies" : {
      "Project" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "type" : "object",
              "properties" : {
                "name" : {
                  "type" : "string",
                  "description" : "Name of project"
                },
                "cwd" : {
                  "type" : "string",
                  "description" : "Absolute project path on server"
                },
                "host" : {
                  "type" : "string",
                  "description" : "Ip address of host machine"
                },
                "user" : {
                  "type" : "string",
                  "description" : "System user which will execute commands"
                }
              }
            }
          }
        }
      },
      "Script" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "type" : "object",
              "properties" : {
                "name" : {
                  "type" : "string",
                  "description" : "Name of script"
                },
                "script" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string",
                    "description" : "Formally lines of script to execute"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}